refactor: extract functions from docs-table
